Revert "requestId when sending a message (#32)"
This reverts commit 210e846adea92a9b97363575aa138bbb7b1c9bf9.
