+++ src/index.ts
@@ -1,23 +1,83 @@
-import {
-  SafeListeners,
-  InterfaceMessageIds,
-  InterfaceMessageEvent,
-  Networks,
-  SDKMessageToPayload,
-  SDKMessageIds,
-  Transaction,
-  SdkInstance,
-  SentSDKMessage,
-  RequestId,
-} from './types';
-import { INTERFACE_MESSAGES, SDK_MESSAGES } from './messageIds';
+/*
+    The reason for duplicating types in both uppercase/lowercase is because in the safe-react
+    type for networks contains uppercase strings and with previous type it resulted in a type error.
+    The sdk converts network to lowercase, so passing an uppercase one is totally valid too.
+*/
+export type Networks =
+  | 'MAINNET'
+  | 'MORDEN'
+  | 'ROPSTEN'
+  | 'RINKEBY'
+  | 'GOERLI'
+  | 'KOVAN'
+  | 'UNKNOWN'
+  | 'mainnet'
+  | 'morden'
+  | 'ropsten'
+  | 'rinkeby'
+  | 'goerli'
+  | 'kovan'
+  | 'unknown';
+
+type ValueOf<T> = T[keyof T];
+
+export interface Transaction {
+  to: string;
+  value: string;
+  data: string;
+}
+
+export interface SdkInstance {
+  addListeners: (listeners: SafeListeners) => void;
+  removeListeners: () => void;
+  sendTransactions: (txs: Transaction[]) => void;
+}
+
+export interface SafeInfo {
+  safeAddress: string;
+  network: Networks;
+  ethBalance: string;
+}
+
+export interface SafeListeners {
+  onSafeInfo: (info: SafeInfo) => void;
+}
+
+interface InterfaceMessageEvent extends MessageEvent {
+  data: {
+    messageId: keyof InterfaceMessages;
+    data: InterfaceMessageToPayload[keyof InterfaceMessages];
+  };
+}
+
+export const SDK_MESSAGES = {
+  SAFE_APP_SDK_INITIALIZED: 'SAFE_APP_SDK_INITIALIZED',
+  SEND_TRANSACTIONS: 'SEND_TRANSACTIONS',
+} as const;
+
+export interface SDKMessageToPayload {
+  [SDK_MESSAGES.SAFE_APP_SDK_INITIALIZED]: undefined;
+  [SDK_MESSAGES.SEND_TRANSACTIONS]: Transaction[];
+}
+
+export type SDKMessages = typeof SDK_MESSAGES;
+
+export const INTERFACE_MESSAGES = {
+  ON_SAFE_INFO: 'ON_SAFE_INFO',
+} as const;
+
+export interface InterfaceMessageToPayload {
+  [INTERFACE_MESSAGES.ON_SAFE_INFO]: SafeInfo;
+}
+
+export type InterfaceMessages = typeof INTERFACE_MESSAGES;
 
 const config: {
   safeAppUrlsRegExp?: RegExp[];
   listeners?: SafeListeners;
 } = {};
 
-const _logMessageFromSafe = (origin: string, messageId: InterfaceMessageIds): void => {
+const _logMessageFromSafe = (origin: string, messageId: ValueOf<InterfaceMessages>): void => {
   console.info(`SafeConnector: A message with id ${messageId} was received from origin ${origin}.`);
 };
 
@@ -63,27 +123,12 @@ const _onParentMessage = async ({ origin, data }: InterfaceMessageEvent): Promis
   }
 };
 
-const _sendMessageToParent = <T extends SDKMessageIds>(
-  messageId: T,
-  data: SDKMessageToPayload[T],
-  requestId?: RequestId,
-): SentSDKMessage<T> => {
-  if (!requestId) {
-    requestId = Math.trunc(window?.performance.now() || Date.now());
-  }
-  const message = {
-    messageId,
-    requestId,
-    data,
-  };
-
-  window.parent.postMessage(message, '*');
-
-  return message;
+const _sendMessageToParent = <T extends keyof SDKMessages>(messageId: T, data?: SDKMessageToPayload[T]): void => {
+  window.parent.postMessage({ messageId, data }, '*');
 };
 
 function sendInitializationMessage(): void {
-  _sendMessageToParent(SDK_MESSAGES.SAFE_APP_SDK_INITIALIZED, undefined);
+  _sendMessageToParent(SDK_MESSAGES.SAFE_APP_SDK_INITIALIZED);
 }
 
 /**
@@ -107,14 +152,11 @@ function removeListeners(): void {
  * Request Safe app to send transactions
  * @param txs
  */
-function sendTransactions(txs: Transaction[], requestId?: RequestId): SentSDKMessage<'SEND_TRANSACTIONS'> {
+function sendTransactions(txs: Transaction[]): void {
   if (!txs || !txs.length) {
-    throw new Error('sendTransactions: No transactions were passed');
+    return;
   }
-
-  const message = _sendMessageToParent(SDK_MESSAGES.SEND_TRANSACTIONS, txs, requestId);
-
-  return message;
+  _sendMessageToParent(SDK_MESSAGES.SEND_TRANSACTIONS, txs);
 }
 
 /**
@@ -133,6 +175,3 @@ function initSdk(safeAppUrlsRegExp: RegExp[] = []): SdkInstance {
 }
 
 export default initSdk;
-
-export * from './types';
-export * from './messageIds';
