--- src/messageIds.ts
@@ -1,8 +1,0 @@
-export const INTERFACE_MESSAGES = {
-  ON_SAFE_INFO: 'ON_SAFE_INFO',
-} as const;
-
-export const SDK_MESSAGES = {
-  SAFE_APP_SDK_INITIALIZED: 'SAFE_APP_SDK_INITIALIZED',
-  SEND_TRANSACTIONS: 'SEND_TRANSACTIONS',
-} as const;
